hash: package for rolling hashes
The LZMA encoder requires to maintain a hash of byte sequences of the
length 2, 4 and maybe 3. For efficient computation rolling hashes are
the right approach. The hash package will at least provide the
Rabin-Karp rolling hash and a Cyclic Polynomial rolling hash.

The package will provide a Roller interface that all rolling hashes will
have to support. The general methods like computing all hashes for a
byte slice will work with that interface.