lzma: better handling of eos marker and rep operation
The commit fixes the long problem of handling the full range of distance
values in repetition operations. In 32-bit systems the int type supports
only 32 bit. So not all possible distance values including the eos
marker can be supported. By using an int64 value this problem is fixed.
Note that the int values have 64-bit by default on 64-bit systems.

This lead to a number of changes and resulted also in a better
documentation.