lzma2: initialize baseReader instead of using a new function
The baseReader will be embedded into the actual readers and will not be
allocated. For this there is a need for an init function that this
commit provideslzma2: initialize baseReader instead of using a new
function

The baseReader will be embedded into the actual readers and will not be
allocated. For this there is a need for an init function that this
commit provides..