lzbase: provided WriterState and ReaderState
With one State type we cannot ensure that a State used for reading will
be combined in a Writer object. Therefore we distinguish between a
ReaderState and a WriterState to ensure that by static typing. The two
seperate type provide also a simple way to access the ReaderDict and
WriterDict implementations stored in the dictionary interface.