lzbase: provided LimitedReader
The LimitedReader provides a LimitedReader for the lzbase.Reader. It
confirms the EOF condition when the limit is reached by checking that
the reader dictionary has been completely read and the range decoder is
drained.

A new test function has been added to writer_test.go that uses the
LimitedReader.