lzma,lzb: fixed buffer size problem
The old buffer size couldn't be given in a reasonable manner as default,
because readers buffer size was taken directly and the writer added
dictionary and buffer size. Now I replaced buffer size with extra buffer
size, which is only required for writing.

That resulted in the introduction of NormalizeReaderSizes and
NormalizeWriterSizes. The change is complete and does apparently work.