lzma: merged lzb into lzma
While implementing the Parameters type for lzma I started to think
about the separation of lzb and lzma and found out that it doesn't make
an awful lot of sense to have to implementation of Parameters. Package
lzma is using practically all functionality of lzb and adds only the
readHeader and writeHeader functions. Since packages lzma and lzb
basically support both the reading and writing of lzma stream I decided
to merge them.

lzma2 will stay a separate package, because it combines LZMA streams.