lzma.Writer: implemented Size method
The Size method returns the number of bytes written to the LZMA stream.
It replaces the N field, which had to be counted separately. This commit
adds a start field that stores the current value of w.buf.top at start.
The method computes then only the difference between w.buf.top and
w.start.