lzma2: resorted tests for cU
The chunk type cU should be larger than cUD as in the actual encoding.
Therefore I resorted the sequence of the tests, which would not be
strictly necessary, but I want the sequence to be obvious to the reader,
because later code relies on the order of the chunk type values.