newlzma: first implementation of the decoder dictionary
This decoder dictionary doesn't implement a separate buffer. At the end
it was simpler to implement the buffer directly in the dictionary. The
WriteRep and WriteByte functions care only for the read length and not
the overall bottom of the pointer. Using also the length of the
dictionary and the read buffer resulted in more readable code.