lzma: changed interface of Matches method
The Matches method returned a slice. This required allocations by the
matcher and didn't allow the the caller to control the maximum number of
positions that should be looked for. By providing the positions slice
as argument to the Matches function, the controller can control the
allocation and the maximum number of positions to look for.