lzma: some minor performance tweaks
This commit contains a range of simple performance tweaks. It appears
that the idea of buffering operations doesn't help with performance. The
speed is quite slow.

A better approach will be a simple greedy match finder looking for the
longest match is much speedier. A simple greedy finder doesn't need an
operations cache. The design will be much simpler and therefore
speedier. On a 3 MHz machine I need to achieve compression speed larger
than 1 MB/s to be acceptable. At the moment I reach only 0.3 MB/s.