xz: support for stream padding
An xz file may consist of more than one XZ stream and between the
streams their may be padding. The readHeader function reads now the
next four bytes and checks whether they are padding and returns an
appropriate function. This way readHeader can be used for searching the
next header.