lzma2: reasonable interfaces for reader and writer
The dictCap must be always given in the default use case for an LZMA2
reader and writer and has therefore been added to the interface. The
reader provides now information about properties value used to encode
the chunks.

The reader structure allows now also to add additional parameters before
the first Read.