lzma: fixed bug with state management
We had a bug where the encoding of a match with length 1 didn't work,
because it didn't match rep0. The cause was that the state had to be
reset, because an uncompressed chunk had to be written in LZMA2.

Since fixing this problem would result in a major redesign of the
compressor, I simply removed the support for length 1 matches.