xz,lzma: moved lzma2 reader and writer to lzma
The split of the lzma and lzma2 package had the problem that it exposed
too many objects that are implementation-internal to the public. This
has been resolved now by moving the lzma2 readers and writers to lzma.
The reader and writer are now named Reader2 and Writer2. All the
exported internal types and functions from lzma are now hidden.