lzma: improved match finding speed (dramatically)
The commit accelerates the compression of files doing two changes. First
we check the byte needed to get a longer match. If this byte doesn't
match, we jump to the next distance candidate. If it matches we check
whether there is indeed a match. We also improved the comparison
function. Both measures halved the time required to compress enwik8. We
have now enough time budget to improve the compression ratio compared to
xz.