xz: add support for check-ID None to Reader
The check-ID None is valid and SHOULD be supported according to the
xz specification. This commit modifies the Reader to support the
check-ID None.

A new simple hash noneHash is introduced, which makes the
implementation quite elegant. For performance reasons we avoid the
writing into the hash during decompression.

Fixes issue #27 - https://github.com/ulikunitz/xz/issues/27