xz: fix bug in readIndexBody
Github user johejo fuzzed xz and found a bug in readIndexBody. The
function reads the record len and allocates the records without further
check. If the number given there is very high the make function panics
because the number of records is too high. This commit provides the
number of expected records to readIndexBody and the function checks the
provided number befor it allocates the records.

Fixes issue #40