lzma: don't use pipes for LZMA2 reader2
A LZMA2 stream needs dictionary resets in regular sequences to be read
in a multithreaded mode. In that case we used a pipe to transfer the
uncompressed data from the worker to the multi-threaded reader. Now we
provide a chunk reader directly to the read function and don't use the
pipe at all.