lzma: fix handling of small dictionary sizes
As Matt Dainty (@bodgit) reported there is an issue if the header of the
LZMA stream is less than the minimum dictionary size of 4096 byte. The
specification of the LZMA format says that in that case a dictionary
size of 4096 byte should be used, our code returns an error.

This commit changes the behavior and adds a simple test case to test for
the right behavior.

Fixes [#52](https://github.com/ulikunitz/xz/pull/52)