Fix bug in multihreaded writing
The Writer benchmark in xz run into a panic, when the block size has
been set to 256 kByte. I found out that the LZMA literal encoder must
have access to the whole window, because it reads a byte which offset to
the window front migt be large. So we have to set ShrinkSize of the
sequence buffer to the window size and need a larger buffer size than
the window. This commit provides the change.