xz.Reader: change semantics of SingleStream flag
The SingleStream flag reads now a single stream and returns EOF. In v0.5
it checked for extra data in the underlying stream, but this can be done
by the client code if required. Implementing the semantics this way may
help two read multiple tar files combined into one file.

This was raised as an issue by LSChyi. Many thanks for reporting the
issue.

Fixed: #54