Remove Type from JSON structures
The Type was intended to have a clear field identifying the type of a
JSON structure. But in practice it is a technicality that makes the use
of the JSON structures inconvenient. I removed the type field.

Additionally the correct format for the LZMA2 reader and writer is now
LZMA2 and not LZMA.