lzma: adapted package to new lz version
lz doesn't export SetDefaults and Verify anymore. This leads to more
clarity what needs to be done here in the lzma package. We are setting
now the WindowSize of the ParserConfig if a WindowSize larger than zero
was provided.

The remainder of the commit is handles the fact that the BufferConfig is
now only available through the ParserConfig.