Fixed chunk writer
The chunk writer had a number of issues related to the handling of the
parser buffer. The buffer must have at least $2^{16}$ byte to support
storing the clear text compression. Pruning needs to maintain all the
data of the current chunk in the buffer until compression is smaller
than the original data.